Added AMBuild support and optimisations

+++ util.cpp
@@ -318,10 +318,10 @@ void UTIL_BuildFileName(char *filename, char *arg1, char *arg2)
 
 void ClampAngle(float &angle)
 {
-   if (angle >= 180)
-      angle -= 360 * static_cast<int>(angle / 360 + 0.5);
-   if (angle < 180)
-      angle += 360 * static_cast<int>(-angle / 360 + 0.5);
+   if (angle >= 180.0f)
+      angle -= 360.0f * (angle / 360.0f + 0.5f);
+   if (angle < 180.0f)
+      angle += 360.0f * (-angle / 360.0f + 0.5f);
 }
 
 void ClampAngles(Vector &angles)
